move native code to its own directory
