+++ flutter/lib/src/integrations/native_app_start_integration.dart
@@ -2,7 +2,7 @@ import 'package:flutter/scheduler.dart';
 import 'package:sentry/sentry.dart';
 
 import '../sentry_flutter_options.dart';
-import '../sentry_native.dart';
+import '../native/sentry_native.dart';
 import '../event_processor/native_app_start_event_processor.dart';
 
 /// Integration which handles communication with native frameworks in order to
