+++ flutter/lib/src/native/method_channel_helper.dart
