+++ flutter/lib/src/native/native_scope_observer.dart
