+++ flutter/test/native_scope_observer_test.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry/sentry.dart';
-import 'package:sentry_flutter/src/native_scope_observer.dart';
+import 'package:sentry_flutter/src/native/native_scope_observer.dart';
 
 import 'mocks.dart';
