+++ flutter/test/sentry_flutter_util.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/file_system_transport.dart';
-import 'package:sentry_flutter/src/native_scope_observer.dart';
+import 'package:sentry_flutter/src/native/native_scope_observer.dart';
 
 void testTransport({
   required Transport transport,
